Address review: move get_worker_info to _worker_info.py
Per reviewer feedback, get_worker_info() and the WorkerInfo setup
logic don't belong in the private _apply_udf.py implementation file.

- New torchdata/nodes/_worker_info.py owns the thread-local storage,
  get_worker_info(), and _set_worker_info() helper
- _apply_udf.py calls _set_worker_info() and is otherwise unchanged
- __init__.py imports get_worker_info from _worker_info
- dataset=None is documented: ParallelMapper maps over items, not
  PyTorch Dataset objects, so there is no dataset to pass
